splitting TokenManager and DjangoTokenManager + tests
