+++ go/conversation/base.py
@@ -18,6 +18,7 @@ from django.http import HttpResponse
 from go.vumitools.conversation.models import (
     CONVERSATION_DRAFT, CONVERSATION_RUNNING, CONVERSATION_FINISHED)
 from go.vumitools.exceptions import ConversationSendError
+from go.base.django_token_manager import DjangoTokenManager
 from go.conversation.forms import (ConversationForm, ConversationGroupForm,
                                     ConfirmConversationForm,
                                     ReplyToMessageForm)
@@ -26,7 +27,6 @@ from go.conversation.tasks import (export_conversation_messages,
 from go.base import message_store_client as ms_client
 from go.base.utils import (make_read_only_form, conversation_or_404,
                             page_range_window)
-from go.base.token_manager import TokenManager
 
 
 class ConversationView(TemplateView):
@@ -219,8 +219,7 @@ class StartConversationView(ConversationView):
         redirect_to = self.get_view_url('confirm',
                             conversation_key=conversation.key)
         # The token to be sent.
-        redis = request.user_api.api.redis
-        token_manager = TokenManager(redis.sub_manager('token_manager'))
+        token_manager = DjangoTokenManager(request.user_api.api.token_manager)
         token = token_manager.generate(redirect_to, user_id=request.user.id,
                                         extra_params=params)
         conversation.send_token_url(token_manager.url_for_token(token),
@@ -250,8 +249,7 @@ class ConfirmConversationView(ConversationView):
                 sending a conversation confirmation SMS will assign a batch key
                 while not actually starting the conversation.
         """
-        redis = request.user_api.api.redis
-        token_manager = TokenManager(redis.sub_manager('token_manager'))
+        token_manager = DjangoTokenManager(request.user_api.api.token_manager)
         token = request.GET.get('token')
         token_data = token_manager.verify_get(token)
         if not token_data:
@@ -264,8 +262,7 @@ class ConfirmConversationView(ConversationView):
 
     def post(self, request, conversation):
         token = request.POST.get('token')
-        redis = request.user_api.api.redis
-        token_manager = TokenManager(redis.sub_manager('token_manager'))
+        token_manager = DjangoTokenManager(request.user_api.api.token_manager)
         token_data = token_manager.verify_get(token)
         if not token_data:
             raise Http404
